requirements.txt
Django==2.2.17
mysqlclient
sqlalchemy
pycryptodome
pymysql
kafka
uwsgi